Delete normalizeSlot, and move the rest of allocation.ts to allocate/storage.ts, adjusting references appropriately.  (Also, fix an incorrect invocation of storageSize in mapping decoding.)
