+++ packages/truffle-decode-utils/src/index.ts
@@ -1,4 +1,3 @@
-export * from "./allocation";
 export * from "./conversion";
 export * from "./evm";
 export * from "./definition";
